refactor: split api/billing into sub-modules and extend BillingRepository
- Split monolithic api/billing/mod.rs (1,066 lines) into focused sub-modules:
- Move 5 admin billing handlers from router.rs to api/admin/billing.rs:
- Extend BillingRepository with 20+ methods:
