+++ src/api/admin/mod.rs
@@ -1,3 +1,4 @@
 pub mod api_keys;
+pub mod billing;
 pub mod blacklist;
 pub mod users;
