+++ src/db/queries.rs
@@ -1,7 +1,6 @@
 use crate::models::{BillingAccount, Link, Organization, User, user::CreateUserData};
 use crate::repositories::link_repository::{AdminLink, serialize_optional_int_as_bool};
 use crate::utils::now_timestamp;
-use wasm_bindgen::JsValue;
 use worker::d1::D1Database;
 use worker::*;
 
@@ -748,558 +747,6 @@ pub async fn get_org_billing_account(db: &D1Database, org_id: &str) -> Result<Op
         Some(val) => Ok(val["billing_account_id"].as_str().map(|s| s.to_string())),
         None => Ok(None),
     }
-}
-
-// ─── Admin Billing Account Queries ────────────────────────────────────────────
-
-/// Response type for billing account list with stats
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct BillingAccountWithStats {
-    pub id: String,
-    pub owner_user_id: String,
-    pub owner_email: String,
-    pub owner_name: Option<String>,
-    pub tier: String,
-    pub org_count: i64,
-    pub total_members: i64,
-    pub links_created_this_month: i64,
-    pub created_at: i64,
-}
-
-/// Response type for org within a billing account
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct OrgWithMembersCount {
-    pub id: String,
-    pub name: String,
-    pub slug: String,
-    pub member_count: i64,
-    pub link_count: i64,
-    pub created_at: i64,
-}
-
-/// Response type for detailed billing account view
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct BillingAccountDetails {
-    pub account: BillingAccount,
-    pub owner: User,
-    pub organizations: Vec<OrgWithMembersCount>,
-    pub usage: UsageStats,
-    pub subscription: Option<serde_json::Value>,
-}
-
-/// Usage stats for billing account
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct UsageStats {
-    pub links_created_this_month: i64,
-    pub max_links_per_month: Option<i64>,
-    pub year_month: String,
-}
-
-/// List all billing accounts with statistics (admin only).
-/// Returns paginated list with aggregated stats.
-pub async fn list_billing_accounts_for_admin(
-    db: &D1Database,
-    page: i64,
-    limit: i64,
-    search: Option<&str>,
-    tier_filter: Option<&str>,
-) -> Result<(Vec<BillingAccountWithStats>, i64)> {
-    let offset = (page - 1) * limit;
-    let current_month = chrono::Utc::now().format("%Y-%m").to_string();
-
-    // Build WHERE clause for filtering
-    let mut where_clauses = vec![];
-    let mut bind_values: Vec<JsValue> = vec![];
-
-    if let Some(search_term) = search {
-        where_clauses.push("u.email LIKE ?");
-        bind_values.push(format!("%{}%", search_term).into());
-    }
-
-    if let Some(tier) = tier_filter {
-        where_clauses.push("ba.tier = ?");
-        bind_values.push(tier.into());
-    }
-
-    let where_sql = if where_clauses.is_empty() {
-        String::new()
-    } else {
-        format!("WHERE {}", where_clauses.join(" AND "))
-    };
-
-    // Count total results
-    let count_sql = format!(
-        "SELECT COUNT(DISTINCT ba.id) as count
-         FROM billing_accounts ba
-         LEFT JOIN users u ON u.id = ba.owner_user_id
-         {}",
-        where_sql
-    );
-    let count_stmt = db.prepare(&count_sql);
-    let count_result = if bind_values.is_empty() {
-        count_stmt.first::<serde_json::Value>(None).await?
-    } else {
-        count_stmt
-            .bind(&bind_values)?
-            .first::<serde_json::Value>(None)
-            .await?
-    };
-    let total = count_result
-        .and_then(|v| v["count"].as_f64())
-        .unwrap_or(0.0) as i64;
-
-    // Fetch paginated results with aggregated stats
-    let mut bind_values_with_limit = bind_values.clone();
-    bind_values_with_limit.push(current_month.clone().into());
-    bind_values_with_limit.push((limit as f64).into());
-    bind_values_with_limit.push((offset as f64).into());
-
-    let query_sql = format!(
-        "SELECT ba.id, ba.owner_user_id, ba.tier, ba.created_at,
-                u.email as owner_email, u.name as owner_name,
-                COUNT(DISTINCT o.id) as org_count,
-                COUNT(DISTINCT om.user_id) as total_members,
-                COALESCE(mc.links_created, 0) as links_created_this_month
-         FROM billing_accounts ba
-         LEFT JOIN users u ON u.id = ba.owner_user_id
-         LEFT JOIN organizations o ON o.billing_account_id = ba.id
-         LEFT JOIN org_members om ON om.org_id = o.id
-         LEFT JOIN monthly_counters mc ON mc.billing_account_id = ba.id AND mc.year_month = ?
-         {}
-         GROUP BY ba.id
-         ORDER BY ba.created_at DESC
-         LIMIT ? OFFSET ?",
-        where_sql
-    );
-
-    let stmt = db.prepare(&query_sql);
-    let results = stmt.bind(&bind_values_with_limit)?.all().await?;
-    let rows = results.results::<serde_json::Value>()?;
-
-    let accounts: Vec<BillingAccountWithStats> = rows
-        .iter()
-        .filter_map(|row| {
-            Some(BillingAccountWithStats {
-                id: row["id"].as_str()?.to_string(),
-                owner_user_id: row["owner_user_id"].as_str()?.to_string(),
-                owner_email: row["owner_email"].as_str()?.to_string(),
-                owner_name: row["owner_name"].as_str().map(|s| s.to_string()),
-                tier: row["tier"].as_str()?.to_string(),
-                org_count: row["org_count"].as_f64()? as i64,
-                total_members: row["total_members"].as_f64()? as i64,
-                links_created_this_month: row["links_created_this_month"].as_f64()? as i64,
-                created_at: row["created_at"].as_f64()? as i64,
-            })
-        })
-        .collect();
-
-    Ok((accounts, total))
-}
-
-/// Get detailed view of a single billing account with all orgs and users.
-pub async fn get_billing_account_details(
-    db: &D1Database,
-    billing_account_id: &str,
-) -> Result<Option<BillingAccountDetails>> {
-    // Get billing account
-    let account = match get_billing_account(db, billing_account_id).await? {
-        Some(acc) => acc,
-        None => return Ok(None),
-    };
-
-    // Get owner user
-    let owner_stmt = db.prepare("SELECT * FROM users WHERE id = ?1");
-    let owner = owner_stmt
-        .bind(&[account.owner_user_id.clone().into()])?
-        .first::<User>(None)
-        .await?;
-
-    let owner = match owner {
-        Some(u) => u,
-        None => return Ok(None),
-    };
-
-    // Get all organizations with member and link counts
-    let orgs_stmt = db.prepare(
-        "SELECT o.id, o.name, o.slug, o.created_at,
-                COUNT(DISTINCT om.user_id) as member_count,
-                COUNT(DISTINCT l.id) as link_count
-         FROM organizations o
-         LEFT JOIN org_members om ON om.org_id = o.id
-         LEFT JOIN links l ON l.org_id = o.id AND l.status = 'active'
-         WHERE o.billing_account_id = ?1
-         GROUP BY o.id
-         ORDER BY o.created_at ASC",
-    );
-    let orgs_result = orgs_stmt.bind(&[billing_account_id.into()])?.all().await?;
-    let orgs_rows = orgs_result.results::<serde_json::Value>()?;
-    let organizations: Vec<OrgWithMembersCount> = orgs_rows
-        .iter()
-        .filter_map(|row| {
-            Some(OrgWithMembersCount {
-                id: row["id"].as_str()?.to_string(),
-                name: row["name"].as_str()?.to_string(),
-                slug: row["slug"].as_str()?.to_string(),
-                member_count: row["member_count"].as_f64()? as i64,
-                link_count: row["link_count"].as_f64()? as i64,
-                created_at: row["created_at"].as_f64()? as i64,
-            })
-        })
-        .collect();
-
-    // Get usage stats for current month
-    let current_month = chrono::Utc::now().format("%Y-%m").to_string();
-    let counter =
-        get_monthly_counter_for_billing_account(db, billing_account_id, &current_month).await?;
-
-    let tier =
-        crate::models::Tier::from_str_value(&account.tier).unwrap_or(crate::models::Tier::Free);
-    let limits = tier.limits();
-
-    let usage = UsageStats {
-        links_created_this_month: counter,
-        max_links_per_month: limits.max_links_per_month,
-        year_month: current_month,
-    };
-
-    // Get subscription info
-    let subscription = get_subscription_for_billing_account(db, billing_account_id).await?;
-
-    Ok(Some(BillingAccountDetails {
-        account,
-        owner,
-        organizations,
-        usage,
-        subscription,
-    }))
-}
-
-/// Update billing account tier (admin only).
-/// This will affect all organizations linked to this billing account.
-pub async fn update_billing_account_tier(
-    db: &D1Database,
-    billing_account_id: &str,
-    new_tier: &str,
-) -> Result<()> {
-    // Update billing account tier
-    let stmt = db.prepare("UPDATE billing_accounts SET tier = ?1 WHERE id = ?2");
-    stmt.bind(&[new_tier.into(), billing_account_id.into()])?
-        .run()
-        .await?;
-
-    Ok(())
-}
-
-/// Update billing account provider customer ID.
-/// Called when a subscription is created to link the billing account to the Polar customer.
-pub async fn update_billing_account_provider_customer_id(
-    db: &D1Database,
-    billing_account_id: &str,
-    provider_customer_id: &str,
-) -> Result<()> {
-    let stmt = db.prepare("UPDATE billing_accounts SET provider_customer_id = ?1 WHERE id = ?2");
-    stmt.bind(&[provider_customer_id.into(), billing_account_id.into()])?
-        .run()
-        .await?;
-    Ok(())
-}
-
-// ─── Billing / Subscription queries ──────────────────────────────────────────
-
-/// Look up a billing_account_id by its provider customer ID.
-/// Used in webhook handlers where external_id is missing from the payload.
-pub async fn get_billing_account_id_by_provider_customer(
-    db: &D1Database,
-    provider_customer_id: &str,
-) -> Result<Option<String>> {
-    let stmt =
-        db.prepare("SELECT id FROM billing_accounts WHERE provider_customer_id = ?1 LIMIT 1");
-    let result = stmt
-        .bind(&[provider_customer_id.into()])?
-        .first::<serde_json::Value>(None)
-        .await?;
-    Ok(result.and_then(|v| v["id"].as_str().map(|s| s.to_string())))
-}
-
-/// Get the current active subscription for a billing account.
-/// Returns the raw JSON row for flexibility.
-pub async fn get_subscription_for_billing_account(
-    db: &D1Database,
-    billing_account_id: &str,
-) -> Result<Option<serde_json::Value>> {
-    let stmt = db.prepare(
-        "SELECT id, billing_account_id, status, plan, interval,
-                provider_subscription_id, provider_customer_id, provider_price_id,
-                current_period_start, current_period_end,
-                cancel_at_period_end, canceled_at, created_at, updated_at,
-                amount_cents, currency, discount_name, pending_cancellation
-         FROM subscriptions
-         WHERE billing_account_id = ?1
-         ORDER BY created_at DESC
-         LIMIT 1",
-    );
-    stmt.bind(&[billing_account_id.into()])?
-        .first::<serde_json::Value>(None)
-        .await
-}
-
-/// Insert or update a subscription record for a billing account.
-/// Uses provider_subscription_id as the unique key for upsert.
-#[allow(clippy::too_many_arguments)]
-pub async fn upsert_subscription(
-    db: &D1Database,
-    billing_account_id: &str,
-    provider_subscription_id: &str,
-    provider_customer_id: &str,
-    status: &str,
-    plan: &str,
-    interval: &str,
-    provider_price_id: &str,
-    current_period_start: i64,
-    current_period_end: i64,
-    cancel_at_period_end: bool,
-    amount_cents: Option<i64>,
-    currency: &str,
-    discount_name: Option<&str>,
-    ends_at: Option<i64>,
-    now: i64,
-) -> Result<()> {
-    let sub_id = format!("sub_{}", crate::utils::generate_short_code_with_length(16));
-    let cancel_flag: i64 = if cancel_at_period_end { 1 } else { 0 };
-
-    let stmt = db.prepare(
-        "INSERT INTO subscriptions (
-           id, billing_account_id, status, plan, interval,
-           provider_subscription_id, provider_customer_id, provider_price_id,
-           current_period_start, current_period_end, ends_at,
-           cancel_at_period_end, amount_cents, currency, discount_name,
-           created_at, updated_at
-         ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14, ?15, ?16, ?16)
-         ON CONFLICT(provider_subscription_id) DO UPDATE SET
-           status = excluded.status,
-           plan = excluded.plan,
-           interval = excluded.interval,
-           provider_price_id = excluded.provider_price_id,
-           current_period_start = excluded.current_period_start,
-           current_period_end = excluded.current_period_end,
-           ends_at = excluded.ends_at,
-           cancel_at_period_end = excluded.cancel_at_period_end,
-           amount_cents = excluded.amount_cents,
-           currency = excluded.currency,
-           discount_name = excluded.discount_name,
-           updated_at = excluded.updated_at",
-    );
-    stmt.bind(&[
-        sub_id.into(),
-        billing_account_id.into(),
-        status.into(),
-        plan.into(),
-        interval.into(),
-        provider_subscription_id.into(),
-        provider_customer_id.into(),
-        provider_price_id.into(),
-        (current_period_start as f64).into(),
-        (current_period_end as f64).into(),
-        ends_at.map(|v| (v as f64).into()).unwrap_or("".into()),
-        (cancel_flag as f64).into(),
-        (amount_cents.unwrap_or(0) as f64).into(),
-        currency.into(),
-        discount_name.unwrap_or("").into(),
-        (now as f64).into(),
-    ])?
-    .run()
-    .await?;
-    Ok(())
-}
-
-/// Mark a subscription as canceled (from subscription.deleted / revoked event).
-pub async fn mark_subscription_canceled(
-    db: &D1Database,
-    provider_subscription_id: &str,
-    now: i64,
-) -> Result<()> {
-    let stmt = db.prepare(
-        "UPDATE subscriptions
-         SET status = 'canceled', canceled_at = ?1, updated_at = ?1
-         WHERE provider_subscription_id = ?2",
-    );
-    stmt.bind(&[(now as f64).into(), provider_subscription_id.into()])?
-        .run()
-        .await?;
-    Ok(())
-}
-
-/// Set subscription as pending cancellation (cancel_at_period_end = true).
-/// The tier will be downgraded when the cron job runs after current_period_end.
-pub async fn set_subscription_pending_cancellation(
-    db: &D1Database,
-    provider_subscription_id: &str,
-    current_period_end: i64,
-) -> Result<()> {
-    let stmt = db.prepare(
-        "UPDATE subscriptions
-         SET pending_cancellation = 1,
-             cancel_at_period_end = 1,
-             updated_at = ?1
-         WHERE provider_subscription_id = ?2",
-    );
-    stmt.bind(&[
-        (current_period_end as f64).into(),
-        provider_subscription_id.into(),
-    ])?
-    .run()
-    .await?;
-    Ok(())
-}
-
-/// Clear pending cancellation flag (e.g., when user uncancels their subscription).
-pub async fn clear_subscription_pending_cancellation(
-    db: &D1Database,
-    provider_subscription_id: &str,
-) -> Result<()> {
-    let stmt = db.prepare(
-        "UPDATE subscriptions
-         SET pending_cancellation = 0,
-             cancel_at_period_end = 0,
-             updated_at = (SELECT strftime('%s', 'now'))
-         WHERE provider_subscription_id = ?1",
-    );
-    stmt.bind(&[provider_subscription_id.into()])?.run().await?;
-    Ok(())
-}
-
-/// Get all subscriptions with pending_cancellation that have expired.
-/// Returns provider_subscription_id and billing_account_id for each.
-pub async fn get_expired_pending_cancellations(
-    db: &D1Database,
-    now: i64,
-) -> Result<Vec<serde_json::Value>> {
-    let stmt = db.prepare(
-        "SELECT provider_subscription_id, billing_account_id, current_period_end
-         FROM subscriptions
-         WHERE pending_cancellation = 1
-           AND current_period_end < ?1
-         LIMIT 1000",
-    );
-    let results = stmt.bind(&[(now as f64).into()])?.all().await?;
-
-    results.results::<serde_json::Value>()
-}
-
-/// Finalize an expired subscription after downgrading the tier.
-pub async fn finalize_expired_subscription(
-    db: &D1Database,
-    provider_subscription_id: &str,
-    now: i64,
-) -> Result<()> {
-    let stmt = db.prepare(
-        "UPDATE subscriptions
-         SET status = 'canceled',
-             pending_cancellation = 0,
-             canceled_at = ?1,
-             updated_at = ?1
-         WHERE provider_subscription_id = ?2",
-    );
-    stmt.bind(&[(now as f64).into(), provider_subscription_id.into()])?
-        .run()
-        .await?;
-    Ok(())
-}
-
-/// Update subscription status (admin only).
-pub async fn update_subscription_status(
-    db: &D1Database,
-    subscription_id: &str,
-    status: &str,
-    now: i64,
-) -> Result<()> {
-    let stmt = db.prepare(
-        "UPDATE subscriptions
-         SET status = ?1, updated_at = ?2
-         WHERE id = ?3",
-    );
-    stmt.bind(&[status.into(), (now as f64).into(), subscription_id.into()])?
-        .run()
-        .await?;
-    Ok(())
-}
-
-/// Get a cached product by price_id
-pub async fn get_cached_product_by_price_id(
-    db: &D1Database,
-    price_id: &str,
-) -> Result<Option<serde_json::Value>> {
-    let stmt = db.prepare(
-        "SELECT id, name, description, price_amount, price_currency,
-                recurring_interval, recurring_interval_count, is_archived,
-                polar_product_id, polar_price_id, created_at, updated_at
-         FROM cached_products
-         WHERE polar_price_id = ?1 AND is_archived = FALSE",
-    );
-
-    let result = stmt
-        .bind(&[price_id.into()])?
-        .first::<serde_json::Value>(None)
-        .await?;
-
-    Ok(result)
-}
-
-// ─── Webhook Idempotency ─────────────────────────────────────────────────────
-
-/// Check if a webhook has already been processed (for idempotency)
-pub async fn webhook_already_processed(
-    db: &D1Database,
-    provider: &str,
-    webhook_id: &str,
-) -> Result<bool> {
-    let stmt = db.prepare(
-        "SELECT 1 FROM processed_webhooks
-         WHERE provider = ?1 AND webhook_id = ?2
-         LIMIT 1",
-    );
-
-    let result = stmt
-        .bind(&[provider.into(), webhook_id.into()])?
-        .first::<serde_json::Value>(None)
-        .await?;
-
-    Ok(result.is_some())
-}
-
-/// Mark a webhook as processed (for idempotency)
-/// ttl_seconds: How long to keep the record (default: 30 days = 2592000 seconds)
-pub async fn mark_webhook_processed(
-    db: &D1Database,
-    provider: &str,
-    webhook_id: &str,
-    event_type: &str,
-    ttl_seconds: i64,
-) -> Result<()> {
-    let now = now_timestamp();
-    let expires_at = now + ttl_seconds;
-
-    let stmt = db.prepare(
-        "INSERT INTO processed_webhooks
-         (id, provider, webhook_id, event_type, processed_at, expires_at)
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-    );
-
-    // Generate unique ID for this record (provider + webhook_id + timestamp)
-    let record_id = format!("{}_{}_{}", provider, webhook_id, now);
-
-    stmt.bind(&[
-        record_id.into(),
-        provider.into(),
-        webhook_id.into(),
-        event_type.into(),
-        (now as f64).into(),
-        (expires_at as f64).into(),
-    ])?
-    .run()
-    .await?;
-
-    Ok(())
 }
 
 #[derive(Debug, serde::Deserialize)]
